Give player context to favorites emit
In a system with a bunch of zones, a change to favorites results in a bunch of "favorites" emits. This change includes the sid (and nts and type) so the developer can at least filter by zone (even though all the favorites are the same.)

This will obviously break existing installs.